requirements.txt
python-dotenv
Pillow
pyobjc